fix(mini): scissor-based tile clipping to remove buffer-zone bleed
Replace coordinate-based Liang-Barsky clipping in LineLayer with
per-tile scissor rect derived from the tile matrix. This is the same
approach MapLibre uses (stencil mask) — no geometry filtering needed,
so real geographic borders at tile boundaries (e.g. Egypt's 22°N line)
are preserved at all zoom levels.

Also:
- Bake 1/4096 into the tile matrix (CPU float64) so shaders use raw
  MVT coords without per-vertex division (avoids mediump precision loss)
- Enable WebGL antialias and OES_element_index_uint (Uint32 indices,
  removes 65535-vertex limit for large polygons)
- Add FillLayer tessellation with earcut + Uint32 index buffers
- Support polygon ring type 3 in LineLayer for country borders
- Add phase5 MapLibre compare demo

+++ src/mini/layers/fill.test.ts
@@ -81,7 +81,7 @@ describe('tessellatePolygon', () => {
     const rings = [[{ x: 0, y: 0 }, { x: 1, y: 0 }, { x: 1, y: 1 }, { x: 0, y: 1 }]]
     const { vertices, indices } = tessellatePolygon(rings)
     expect(vertices).toBeInstanceOf(Float32Array)
-    expect(indices).toBeInstanceOf(Uint16Array)
+    expect(indices).toBeInstanceOf(Uint32Array)
     expect(indices.length).toBe(6)
   })
